Cosmetic:fix typos and warnings

+++ src/subtitles/VobSubFile.cpp
@@ -1802,7 +1802,7 @@ bool CVobSubFile::SaveScenarist(CString fn)
 	{
         sizeof(BITMAPINFOHEADER),
 		720, m_size.cy-2, 1, 4, 0,
-		360*(m_size.cy-2),
+		DWORD(360*(m_size.cy-2)),
 		0, 0,
 		16, 4
 	};
@@ -2032,7 +2032,7 @@ bool CVobSubFile::SaveMaestro(CString fn)
 	{
         sizeof(BITMAPINFOHEADER),
 		720, m_size.cy-2, 1, 4, 0,
-		360*(m_size.cy-2),
+		DWORD(360*(m_size.cy-2)),
 		0, 0,
 		16, 4
 	};
